pyproject.toml
[tool.poetry]
name = "engage-engine"
version = "0.0.1"
description = ""
authors = ["BD103 <dont@stalk.me>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2" # Commands
toml = "^0.10.2" # Config
aiohttp = "^3.7.3" # Back-end
python-socketio = "^5.1.0" # Chat Channel
jinja2 = "^2.11.3" # Template Formatting

[tool.poetry.dev-dependencies]
isort = "^5.8.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
